added angles

+++ include/ArmClaw.h
@@ -10,42 +10,42 @@ using namespace std;
 
 class ArmClaw : public ArmPart {
     private:
-        static const double validateClawAngle(const double angle);
-        const double getClawRad(const double angle);
-        const double getClawAngleFromRad(const double rad);
+        static const float validateClawAngle(const float angle);
+        const float getClawRad(const float angle);
+        const float getClawAngleFromRad(const float rad);
     public:    
-        double clawRad = 0;
-        ArmClaw(ArmWrist wrist, const double xAngle, const double yAngle, const double zAngle, const double clawAngle);
-        const double getAngle(const bool validate = true);        
+        float clawRad = 0;
+        ArmClaw(ArmWrist wrist, const float xAngle, const float yAngle, const float zAngle, const float clawAngle);
+        const float getAngle(const bool validate = true);        
         
         virtual const String name() override {
             return "Claw";
         };                
-        const double getXLength() override {
+        const float getXLength() override {
             return CLAW_X_LENGTH;
         };        
-        const double getYLength() override {
+        const float getYLength() override {
             return CLAW_Y_LENGTH;
         };        
-        const double getZLength() override {
+        const float getZLength() override {
             return CLAW_Z_LENGTH;
         };
-        const double getXBase() override {
+        const float getXBase() override {
             return CLAW_X_BASE;
         };
-        const double getYBase() override {
+        const float getYBase() override {
             return CLAW_Y_BASE;
         };
-        const double getZBase() override {
+        const float getZBase() override {
             return CLAW_Z_BASE;
         };
-        const double getWidth() override {
+        const float getWidth() override {
             return CLAW_WIDTH;
         };
-        const double getXScale() override {
+        const float getXScale() override {
             return CLAW_X_SCALE;
         };
-        const double getZScale() override {
+        const float getZScale() override {
             return CLAW_Z_SCALE;
         };
 };
