+++ include/ArmWrist.h
@@ -9,23 +9,23 @@
 
 class ArmWrist : public ArmPart {
     public: 
-        ArmWrist(ArmElbow elbow, const double yAngle);
+        ArmWrist(ArmElbow elbow, const float yAngle);
         virtual const String name() override {
             return "Wrist";
         };        
-        virtual const double getYLength() override {
+        virtual const float getYLength() override {
             return WRIST_Y_LENGTH;
         };
-        virtual const double getYBase() override {
+        virtual const float getYBase() override {
             return WRIST_Y_BASE;
         };
-        virtual const double getYScale() override {
+        virtual const float getYScale() override {
             return WRIST_Y_SCALE;
         };
-        virtual const double getYMaxAngle() override {
+        virtual const float getYMaxAngle() override {
             return WRIST_Y_MAX;
         };
-        virtual const double getYMinAngle() override {
+        virtual const float getYMinAngle() override {
             return WRIST_Y_MIN;
         };
