+++ src/ArmBase.cpp
@@ -2,17 +2,17 @@
 
 //------Coords------
 
-Coords::Coords(const double length, const double YRad, const double ZRad) : length(length),
+Coords::Coords(const float length, const float YRad, const float ZRad) : length(length),
   x(ArmBase::getXLength(length, YRad, ZRad)),
   y(ArmBase::getYLength(length, YRad, ZRad)),
   z(ArmBase::getZLength(length, YRad, ZRad))
 {}
 
-const bool Coords::isEqual(const double x, const double y, const double z, const double tolerance)
+const bool Coords::isEqual(const float x, const float y, const float z, const float tolerance)
 {
-  const double deltaX = this->x - x;
-  const double deltaY = this->y - y;
-  const double deltaZ = this->z - z;
+  const float deltaX = this->x - x;
+  const float deltaY = this->y - y;
+  const float deltaZ = this->z - z;
   return (deltaX < tolerance) && (deltaX > -tolerance) &&
          (deltaY < tolerance) && (deltaY > -tolerance) &&
          (deltaZ < tolerance) && (deltaZ > -tolerance);
@@ -22,20 +22,20 @@ const bool Coords::isEqual(const double x, const double y, const double z, const
 
 void ArmBase::updateCoords()
 {
-  const double zLength = getZLength();  
-  const double zxLocal = zLength * cos(ZRadLocal);
-  const double zyLocal = zLength * sin(ZRadLocal);
-  const double zzLocal = 0;
+  const float zLength = getZLength();  
+  const float zxLocal = zLength * cos(ZRadLocal);
+  const float zyLocal = zLength * sin(ZRadLocal);
+  const float zzLocal = 0;
   
-  const double yLength = getYLength();  
-  const double yxLocal = yLength * cos(YRadLocal) * cos(ZRadLocal);
-  const double yyLocal = yLength * cos(YRadLocal) * sin(ZRadLocal);
-  const double yzLocal = yLength * sin(YRadLocal);
+  const float yLength = getYLength();  
+  const float yxLocal = yLength * cos(YRadLocal) * cos(ZRadLocal);
+  const float yyLocal = yLength * cos(YRadLocal) * sin(ZRadLocal);
+  const float yzLocal = yLength * sin(YRadLocal);
   
-  const double xLength = getXLength();  
-  const double xxLocal = xLength * cos(YRadLocal) * cos(ZRadLocal);
-  const double xyLocal = xLength * cos(YRadLocal) * sin(ZRadLocal);
-  const double xzLocal = xLength * sin(YRadLocal);
+  const float xLength = getXLength();  
+  const float xxLocal = xLength * cos(YRadLocal) * cos(ZRadLocal);
+  const float xyLocal = xLength * cos(YRadLocal) * sin(ZRadLocal);
+  const float xzLocal = xLength * sin(YRadLocal);
 
   /*Serial.printf("\nname: %s\n", name().c_str());
   Serial.printf("ZRadLocal: %f, YRadLocal: %f, XRadLocal: %f\n", ZRadLocal, YRadLocal, ZRadLocal);
@@ -56,164 +56,164 @@ void ArmBase::updateCoords()
   //Serial.printf("rotate z x: %f, y: %f, z: %f\n", this->x, this->y, this->z);  
 }
 
-void ArmBase::rotateZ(const double zRad) { 
-  const double sx = this->x;  
-  const double sy = this->y;
-  const double sz = this->z;    
+void ArmBase::rotateZ(const float zRad) { 
+  const float sx = this->x;  
+  const float sy = this->y;
+  const float sz = this->z;    
   this->x = sx * cos(zRad) - sy * sin(zRad);
   this->y = sx * sin(zRad) + sy * cos(zRad);
   this->z = sz;
 }
 
-void ArmBase::rotateX(const double xRad) { 
-  const double sx = this->x;
-  const double sy = this->y;
-  const double sz = this->z;    
+void ArmBase::rotateX(const float xRad) { 
+  const float sx = this->x;
+  const float sy = this->y;
+  const float sz = this->z;    
   this->x = sx;
   this->y = sy * cos(xRad) - sz * sin(xRad);
   this->z = sy * sin(xRad) + sz * cos(xRad);
   
 }
 
-void ArmBase::rotateY(const double yRad) { 
-  const double sx = this->x;
-  const double sy = this->y;
-  const double sz = this->z;    
+void ArmBase::rotateY(const float yRad) { 
+  const float sx = this->x;
+  const float sy = this->y;
+  const float sz = this->z;    
   this->x = sx * cos(yRad) - sz * sin(yRad);
   this->y = sy;
   this->z = sx * sin(yRad) + sz * cos(yRad);
 }
 
-void ArmBase::setRads(const double xRad, const double yRad, const double zRad)
+void ArmBase::setRads(const float xRad, const float yRad, const float zRad)
 {
   this->YRad = yRad;
   this->ZRad = zRad;
   this->XRad = xRad;
 }
 
-void ArmBase::setRadsLocal(const double xRad, const double yRad, const double zRad)
+void ArmBase::setRadsLocal(const float xRad, const float yRad, const float zRad)
 {
   this->XRadLocal = xRad;
   this->YRadLocal = yRad;
   this->ZRadLocal = zRad;
 }
 
-const double ArmBase::getXRad(const double angle)
+const float ArmBase::getXRad(const float angle)
 {
   return (this->getXScale() * angle + this->getXBase()) * PI / 180.0;
 }
 
-const double ArmBase::getYRad(const double angle)
+const float ArmBase::getYRad(const float angle)
 {
   return (this->getYScale() * angle + this->getYBase()) * PI / 180.0;
 }
 
-const double ArmBase::getZRad(const double angle)
+const float ArmBase::getZRad(const float angle)
 {
   return (this->getZScale() * angle + this->getZBase()) * PI / 180.0;
 }
 
-const double ArmBase::getXRad()
+const float ArmBase::getXRad()
 {
   return this->XRad + this->XRadLocal;
 }
 
-const double ArmBase::getYRad()
+const float ArmBase::getYRad()
 {
   return this->YRad + this->YRadLocal;
 }
 
-const double ArmBase::getZRad()
+const float ArmBase::getZRad()
 {
   return this->ZRad + this->ZRadLocal;
 }
 
-const double ArmBase::getXAngleFromRad(const double rad)
+const float ArmBase::getXAngleFromRad(const float rad)
 {
   if (rad == NAN)
     return NAN;
-  const double sdeg = rad / PI * 180.0;
-  double sAngle = (sdeg - this->getXBase()) / this->getXScale();
+  const float sdeg = rad / PI * 180.0;
+  float sAngle = (sdeg - this->getXBase()) / this->getXScale();
   if (sAngle < 0)
     sAngle = 360 + sAngle;
-  const double tAngle = sAngle - (trunc(sAngle / 360) * 360);
+  const float tAngle = sAngle - (trunc(sAngle / 360) * 360);
   return tAngle;
 }
 
-const double ArmBase::getYAngleFromRad(const double rad)
+const float ArmBase::getYAngleFromRad(const float rad)
 {
   if (rad == NAN)
     return NAN;
-  const double sdeg = rad / PI * 180.0;
-  double sAngle = (sdeg - this->getYBase()) / this->getYScale();
+  const float sdeg = rad / PI * 180.0;
+  float sAngle = (sdeg - this->getYBase()) / this->getYScale();
   if (sAngle < 0)
     sAngle = 360 + sAngle;
-  const double tAngle = sAngle - (trunc(sAngle / 360) * 360);
+  const float tAngle = sAngle - (trunc(sAngle / 360) * 360);
   return tAngle;
 }
 
-const double ArmBase::getZAngleFromRad(const double rad)
+const float ArmBase::getZAngleFromRad(const float rad)
 {
   if (rad == NAN)
     return NAN;
-  const double sdeg = rad / PI * 180.0;
-  double sAngle = (sdeg - this->getZBase()) / this->getZScale();
+  const float sdeg = rad / PI * 180.0;
+  float sAngle = (sdeg - this->getZBase()) / this->getZScale();
   if (sAngle < 0)
     sAngle = 360 + sAngle;
-  const double tAngle = sAngle - (trunc(sAngle / 360) * 360);
+  const float tAngle = sAngle - (trunc(sAngle / 360) * 360);
   return tAngle;
 }
 
-const double ArmBase::getRadFromXY(const double x, const double y)
+const float ArmBase::getRadFromXY(const float x, const float y)
 {
-  const double lengthV = sqrt(x * x + y * y);
-  const double sr = (lengthV != 0) ? y / lengthV : 0;
-  const double siny = asin(sr);
-  const double res = (x < 0) ? (y < 0) ? -PI - siny : PI - siny
+  const float lengthV = sqrt(x * x + y * y);
+  const float sr = (lengthV != 0) ? y / lengthV : 0;
+  const float siny = asin(sr);
+  const float res = (x < 0) ? (y < 0) ? -PI - siny : PI - siny
                              : siny;
   return res;
 }
 
-const double ArmBase::getYRadFromPos(const double localX, const double localY, const double localZ, const double length)
+const float ArmBase::getYRadFromPos(const float localX, const float localY, const float localZ, const float length)
 {
-  const double tLength = isnan(length) ? getLength() : length;
+  const float tLength = isnan(length) ? getLength() : length;
   Serial.printf("ZRad: %f, localZ: %f, tLength: %f\n", ZRad, localZ, tLength);
-  const double sinz = (tLength > 0) ? localZ / tLength : 0;
-  const double sina = asin(sinz);
-  const double asina = abs(sina);
+  const float sinz = (tLength > 0) ? localZ / tLength : 0;
+  const float sina = asin(sinz);
+  const float asina = abs(sina);
   Serial.printf("sinz: %f, asinz: %f\n", sinz, sina);
   Coords coord = Coords(tLength, sina, ZRad);
   Serial.printf("coord.x: %f, coord.y: %f, coord.z: %f\n", coord.x, coord.y, coord.z);
   Serial.printf("localX: %f, localY: %f, localZ: %f\n", localX, localY, localZ);
   if (coord.isEqual(localX, localY, localZ))
     return sina;
-  const double a1 = PI + asina;
+  const float a1 = PI + asina;
   Coords coord2 = Coords(tLength, a1, ZRad);
   if (coord2.isEqual(localX, localY, localZ))
     return a1;
-  const double a2 = PI - asina;
+  const float a2 = PI - asina;
   Coords coord3 = Coords(tLength, a2, ZRad);
   if (coord3.isEqual(localX, localY, localZ))
     return a2;
   return NAN;
 }
 
-const bool ArmBase::isEqual(const double op1, const double op2, const double tolerance)
+const bool ArmBase::isEqual(const float op1, const float op2, const float tolerance)
 {
-  const double delta = op1 - op2;
+  const float delta = op1 - op2;
   return (delta < tolerance) && (delta > -tolerance);
 }
 
-const double ArmBase::getLengthFromPoint(const double x, const double y, const double z)
+const float ArmBase::getLengthFromPoint(const float x, const float y, const float z)
 {
-  const double dz = z;
-  const double qz = dz * dz;
-  const double qx = x * x;
-  const double qy = y * y;
+  const float dz = z;
+  const float qz = dz * dz;
+  const float qx = x * x;
+  const float qy = y * y;
   return sqrt(qx + qy + qz);
 }
 
-const double ArmBase::validateXAngle(const double angle)
+const float ArmBase::validateXAngle(const float angle)
 {
   if (isnan(angle))
     return ERROR_X_ANGLE_IS_NAN;
@@ -224,7 +224,7 @@ const double ArmBase::validateXAngle(const double angle)
   return angle;
 }
 
-const double ArmBase::validateYAngle(const double angle)
+const float ArmBase::validateYAngle(const float angle)
 {
   if (isnan(angle))
     return ERROR_Y_ANGLE_IS_NAN;
@@ -235,7 +235,7 @@ const double ArmBase::validateYAngle(const double angle)
   return angle;
 }
 
-const double ArmBase::validateZAngle(const double angle)
+const float ArmBase::validateZAngle(const float angle)
 {
   if (isnan(angle))
     return ERROR_Z_ANGLE_IS_NAN;
@@ -246,25 +246,25 @@ const double ArmBase::validateZAngle(const double angle)
   return angle;
 }
 
-const double ArmBase::getXAngle(const bool validate)
+const float ArmBase::getXAngle(const bool validate)
 {
-  const double angle = getXAngleFromRad(XRadLocal);
+  const float angle = getXAngleFromRad(XRadLocal);
   if (!validate)
     return angle;
   return ArmBase::validateXAngle(angle);
 }
 
-const double ArmBase::getYAngle(const bool validate)
+const float ArmBase::getYAngle(const bool validate)
 {
-  const double angle = getYAngleFromRad(YRadLocal);
+  const float angle = getYAngleFromRad(YRadLocal);
   if (!validate)
     return angle;
   return ArmBase::validateYAngle(angle);
 }
 
-const double ArmBase::getZAngle(const bool validate)
+const float ArmBase::getZAngle(const bool validate)
 {
-  const double angle = getZAngleFromRad(ZRadLocal);
+  const float angle = getZAngleFromRad(ZRadLocal);
   if (!validate)
     return angle;
   return ArmBase::validateZAngle(angle);
@@ -274,38 +274,38 @@ const bool ArmBase::isValid()
 {
   if (isnan(XRad) || isnan(YRad) || isnan(ZRad))
     return false;
-  const double sMinXRad = getXRad(getXMinAngle());
-  const double sMaxXRad = getXRad(getXMaxAngle());
-  const double minXRad = (sMinXRad > sMaxXRad) ? sMaxXRad : sMinXRad;
-  const double maxXRad = (sMinXRad > sMaxXRad) ? sMinXRad : sMaxXRad;
+  const float sMinXRad = getXRad(getXMinAngle());
+  const float sMaxXRad = getXRad(getXMaxAngle());
+  const float minXRad = (sMinXRad > sMaxXRad) ? sMaxXRad : sMinXRad;
+  const float maxXRad = (sMinXRad > sMaxXRad) ? sMinXRad : sMaxXRad;
   if ((XRad < minXRad) || (XRad > maxXRad))
     return false;
-  const double sMinYRad = getYRad(getYMinAngle());
-  const double sMaxYRad = getYRad(getYMaxAngle());
-  const double minYRad = (sMinYRad > sMaxYRad) ? sMaxYRad : sMinYRad;
-  const double maxYRad = (sMinYRad > sMaxYRad) ? sMinYRad : sMaxYRad;
+  const float sMinYRad = getYRad(getYMinAngle());
+  const float sMaxYRad = getYRad(getYMaxAngle());
+  const float minYRad = (sMinYRad > sMaxYRad) ? sMaxYRad : sMinYRad;
+  const float maxYRad = (sMinYRad > sMaxYRad) ? sMinYRad : sMaxYRad;
   if ((YRad < minYRad) || (YRad > maxYRad))
     return false;
-  const double sMinZRad = getZRad(getZMinAngle());
-  const double sMaxZRad = getZRad(getZMaxAngle());
-  const double minZRad = (sMinZRad > sMaxZRad) ? sMaxZRad : sMinZRad;
-  const double maxZRad = (sMinZRad > sMaxZRad) ? sMinZRad : sMaxZRad;
+  const float sMinZRad = getZRad(getZMinAngle());
+  const float sMaxZRad = getZRad(getZMaxAngle());
+  const float minZRad = (sMinZRad > sMaxZRad) ? sMaxZRad : sMinZRad;
+  const float maxZRad = (sMinZRad > sMaxZRad) ? sMinZRad : sMaxZRad;
   if ((ZRad < minZRad) || (ZRad > maxZRad))
     return false;
   return true;
 }
 
-const double ArmBase::getXLength(const double length, const double YRad, const double ZRad)
+const float ArmBase::getXLength(const float length, const float YRad, const float ZRad)
 {
   return length * cos(YRad) * cos(ZRad);
 }
 
-const double ArmBase::getYLength(const double length, const double YRad, const double ZRad)
+const float ArmBase::getYLength(const float length, const float YRad, const float ZRad)
 {
   return length * cos(YRad) * sin(ZRad);
 }
 
-const double ArmBase::getZLength(const double length, const double YRad, const double ZRad)
+const float ArmBase::getZLength(const float length, const float YRad, const float ZRad)
 {  
   return length * sin(YRad);
 }
