+++ src/ArmPart.cpp
@@ -1,60 +1,60 @@
 #include <ArmPart.h>
 
-void ArmPart::build(ArmBase &base, const double xAngle, const double yAngle, const double zAngle)
+void ArmPart::build(ArmBase &base, const float xAngle, const float yAngle, const float zAngle)
 {
-  const double xRad = getXRad(xAngle);
-  const double yRad = getYRad(yAngle);
-  const double zRad = getZRad(zAngle);
+  const float xRad = getXRad(xAngle);
+  const float yRad = getYRad(yAngle);
+  const float zRad = getZRad(zAngle);
   setRadsLocal(xRad, yRad, zRad);
-  const double pxRad = base.getXRad();
-  const double pyRad = base.getYRad();
-  const double pzRad = base.getZRad();
+  const float pxRad = base.getXRad();
+  const float pyRad = base.getYRad();
+  const float pzRad = base.getZRad();
   setRads(pxRad, pyRad, pzRad);
   updateCoords();
 }
 
-const double ArmPart::getPointLength(ArmBase *prev, const double x, const double y, const double z)
+const float ArmPart::getPointLength(ArmBase *prev, const float x, const float y, const float z)
 {
-  const double px = (prev == NULL) ? 0 : prev->x;
-  const double py = (prev == NULL) ? 0 : prev->y;
-  const double pz = (prev == NULL) ? 0 : prev->z;
-  const double sx = x - px;
-  const double sy = y - py;
-  const double sz = z - pz;
-  const double qx = sx * sx;
-  const double qy = sy * sy;
-  const double qz = sz * sz;
+  const float px = (prev == NULL) ? 0 : prev->x;
+  const float py = (prev == NULL) ? 0 : prev->y;
+  const float pz = (prev == NULL) ? 0 : prev->z;
+  const float sx = x - px;
+  const float sy = y - py;
+  const float sz = z - pz;
+  const float qx = sx * sx;
+  const float qy = sy * sy;
+  const float qz = sz * sz;
   return sqrt(qx + qy + qz);
 }
 
-void ArmPart::setPos(ArmBase *prev, const double x, const double y, const double z)
+void ArmPart::setPos(ArmBase *prev, const float x, const float y, const float z)
 {
-  const double px = (prev == NULL) ? 0 : prev->x;
-  const double py = (prev == NULL) ? 0 : prev->y;
-  const double pz = (prev == NULL) ? 0 : prev->z;
-  const double xrad = (prev == NULL) ? 0 : prev->XRad;
-  const double pzRad = (prev == NULL) ? 0 : prev->ZRad;
-  const double localX = x - px;
-  const double localY = y - py;
-  const double localZ = z - pz;
-  const double zrad = ArmBase::getRadFromXY(x, y);
-  const double yrad = getYRadFromPos(localX, localY, localZ);
+  const float px = (prev == NULL) ? 0 : prev->x;
+  const float py = (prev == NULL) ? 0 : prev->y;
+  const float pz = (prev == NULL) ? 0 : prev->z;
+  const float xrad = (prev == NULL) ? 0 : prev->XRad;
+  const float pzRad = (prev == NULL) ? 0 : prev->ZRad;
+  const float localX = x - px;
+  const float localY = y - py;
+  const float localZ = z - pz;
+  const float zrad = ArmBase::getRadFromXY(x, y);
+  const float yrad = getYRadFromPos(localX, localY, localZ);
   Serial.printf("yRad: %f, zRad: %f\n", yrad, zrad);
 
   //setRads(xrad, yrad, pzRad);
 
-  /*const double xLength = getXLength();
-    const double yLength = getYLength();
-    const double zLength = getZLength();
-    const double xyLength = xLength + yLength;
+  /*const float xLength = getXLength();
+    const float yLength = getYLength();
+    const float zLength = getZLength();
+    const float xyLength = xLength + yLength;
     Coords xCoord = Coords{ xyLength, yRad, zRad };
-    const double cxx = x - xCoord.x;
-    const double cyx = x - xCoord.y;
-    const double czx = x - xCoord.z;
+    const float cxx = x - xCoord.x;
+    const float cyx = x - xCoord.y;
+    const float czx = x - xCoord.z;
     Serial.printf("cxx: %f, cyx: %f, czx: %f\n", cxx, cyx, czx);*/
 }
 
-const double ArmPart::getMaxHorzLength()
+const float ArmPart::getMaxHorzLength()
 {      
   return 0;
 }
